chore: move tests file to tests folder
