+++ indexer/tests/constants.ts
@@ -1,7 +1,7 @@
 // Path to the JSON file containing real transaction data
 export const TRANSACTIONS_DATA_FILE =
-  "indexer/src/test-data/transactionsData.json";
+  "indexer/tests/test-data/transactionsData.json";
 
 // Path to the JSON file containing the expected data trasformed with transform function and toTypedEthTx
 export const EXPECTED_TRANSFORM_DATA_FILE =
-  "indexer/src/test-data/expectedTransformData.json";
+  "indexer/tests/test-data/expectedTransformData.json";
