+++ indexer/tests/header.test.ts
@@ -4,14 +4,14 @@ import {
   Bloom,
   bytesToHex,
   PrefixedHexString,
-} from "../deps.ts";
+} from "../src/deps.ts";
 import { assertEquals } from "https://deno.land/std@0.213.0/assert/assert_equals.ts";
 import {
   DEFAULT_BLOCK_GAS_LIMIT,
   JsonRpcBlock,
   toEthHeader,
-} from "./header.ts";
-import { padString } from "../utils/hex.ts";
+} from "../src/types/header.ts";
+import { padString } from "../src/utils/hex.ts";
 import sinon from "npm:sinon";
 import { KAKAROT } from "../provider.ts";
 import { NULL_HASH } from "../constants.ts";
